+++ HISTORY.md
@@ -14,6 +14,9 @@
 * **Translation Updates:**
     * Russian, thanks, Pavel!
 
+### 2020-10-22
+* fixed UNDO/REDO for "extracted" (single) command blocks
+
 ### 2020-10-21
 * gui: wait until all costumes have loaded before auto-triggering the green-flag event 
 * gui, objects, store: wait until all sounds have loaded before auto-triggering the green-flag event